KTNB-1201: Rename Notebook Manipulator to Notekit across all modules
