+++ .junie/guidelines/junie.guidelines
@@ -1,5 +1,5 @@
 Guidelines to edit widgets project are stored here:
 integrations/widgets/AGENTS.md
 
-Guidelines to edit notebook-manipulator project are stored here:
-integrations/notebook-manipulator/AGENTS.md
+Guidelines to edit notekit project are stored here:
+integrations/notekit/AGENTS.md
