--- integrations/notebook-manipulator/notebook-manipulator-api/src/main/kotlin/org/jetbrains/kotlinx/jupyter/notebook/NotebookManipulatorException.kt
@@ -1,10 +1,0 @@
-package org.jetbrains.kotlinx.jupyter.notebook
-
-/**
- * Exception thrown when a notebook manipulation operation fails.
- */
-public class NotebookManipulatorException(
-    message: String,
-    public val errorCode: String? = null,
-    cause: Throwable? = null,
-) : Exception(message, cause)
