+++ integrations/notekit/build.gradle.kts
