+++ integrations/notekit/notekit-api/build.gradle.kts
@@ -1,0 +1,35 @@
+plugins {
+    alias(libs.plugins.kotlin.jvm)
+    alias(libs.plugins.publisher)
+    alias(libs.plugins.kotlin.serialization)
+    alias(libs.plugins.kotlin.jupyter.api)
+}
+
+dependencies {
+    api(projects.integrations.util.notebookParser)
+    api(libs.kotlinx.serialization.json)
+    api(libs.kotlinx.coroutines.core)
+
+    testImplementation(libs.kotlin.test)
+    testImplementation(libs.test.kotlintest.assertions)
+    testImplementation(libs.kotlinx.serialization.json)
+}
+
+tasks.test {
+    useJUnitPlatform()
+}
+
+kotlin {
+    jvmToolchain(
+        libs.versions.jvm.toolchain
+            .get()
+            .toInt(),
+    )
+    explicitApi()
+}
+
+kotlinPublications {
+    publication {
+        description.set("Kotlin APIs for Jupyter Notebook with Notekit")
+    }
+}
