+++ integrations/notekit/notekit-api/src/main/kotlin/org/jetbrains/kotlinx/jupyter/notebook/NotekitException.kt
@@ -1,0 +1,10 @@
+package org.jetbrains.kotlinx.jupyter.notebook
+
+/**
+ * Exception thrown when a notebook operation fails.
+ */
+public class NotekitException(
+    message: String,
+    public val errorCode: String? = null,
+    cause: Throwable? = null,
+) : Exception(message, cause)
