+++ integrations/notekit/notekit-api/src/test/kotlin/org/jetbrains/kotlinx/jupyter/notebook/JsonConversionsTest.kt
